+++ roqoqo-qasm/tests/integration/backend.rs
@@ -12,13 +12,14 @@
 //
 //! Testing the roqoqo-qasm Backend
 
-use roqoqo::prelude::*;
-use roqoqo::{operations::*, Circuit};
-use roqoqo_qasm::Backend;
-// use roqoqo_test::prepare_monte_carlo_gate_test;
 use std::env::temp_dir;
 use std::fs;
 use std::path::Path;
+
+use roqoqo::prelude::*;
+use roqoqo::{operations::*, Circuit};
+use roqoqo_qasm::Backend;
+
 use test_case::test_case;
 
 /// Test simple circuit with a Definition, a GateOperation and a PragmaOperation
@@ -46,7 +47,7 @@ fn run_simple_circuit(qasm_version: &str, qubits: &str, bits: &str) {
     } else {
         "ctrl @ x"
     };
-    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n{bits};\nrx(1.5707963267948966) qr[0];\nx qr[1];\nmeasure qr -> ro;\n");
+    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n\n{bits};\nrx(1.5707963267948966) qr[0];\nx qr[1];\nmeasure qr -> ro;\n");
     file_name.push_str(".qasm");
     let read_in_path = temp_dir().join(Path::new(file_name.as_str()));
     let extracted = fs::read_to_string(&read_in_path);
@@ -79,7 +80,7 @@ fn simple_circuit_iterator_to_file(qasm_version: &str, qubits: &str, bits: &str)
     } else {
         "ctrl @ x"
     };
-    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n{bits};\nrx(1.5707963267948966) q[0];\nx q[1];\nmeasure q -> ro;\n");
+    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n\n{bits};\nrx(1.5707963267948966) q[0];\nx q[1];\nmeasure q -> ro;\n");
     file_name.push_str(".qasm");
     let read_in_path = temp_dir().join(Path::new(file_name.as_str()));
     let extracted = fs::read_to_string(&read_in_path);
@@ -104,7 +105,7 @@ fn test_duplicate_definitions(qasm_version: &str, qubits: &str, bits: &str) {
     } else {
         "ctrl @ x"
     };
-    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n{bits};\nx qr[0];\nx qr[1];\nmeasure qr -> ro;\n");
+    let lines = format!("OPENQASM {qasm_version};\n\ngate u3(theta,phi,lambda) q {{ U(theta,phi,lambda) q; }}\ngate u2(phi,lambda) q {{ U(pi/2,phi,lambda) q; }}\ngate u1(lambda) q {{ U(0,0,lambda) q; }}\ngate rx(theta) a {{ u3(theta,-pi/2,pi/2) a; }}\ngate ry(theta) a {{ u3(theta,0,0) a; }}\ngate rz(phi) a {{ u1(phi) a; }}\ngate cx c,t {{ {cnot} c,t; }}\n\ngate x a {{ u3(pi,0,pi) a; }}\n\n{qubits};\n\n{bits};\nx qr[0];\nx qr[1];\nmeasure qr -> ro;\n");
     assert_eq!(output, lines);
 }
 
@@ -157,4 +158,26 @@ fn test_debug_clone_partialeq() {
     assert!(backend == backend_0);
     assert!(backend_2 != backend);
     assert!(backend != backend_2);
+}
+
+#[test]
+#[cfg(feature = "unstable_qasm_import")]
+fn test_parsing_methods() {
+    use std::fs::File;
+    use std::io::BufRead;
+    use std::io::BufReader;
+
+    let backend = Backend::new(None, None).unwrap();
+
+    let file = File::open(std::env::current_dir().unwrap().join("tests/input.qasm")).unwrap();
+    let result_from_file = backend.file_to_circuit(file);
+    assert!(result_from_file.is_ok());
+
+    let file = File::open(std::env::current_dir().unwrap().join("tests/input.qasm")).unwrap();
+    let unparsed_file = BufReader::new(file)
+        .lines()
+        .map(|line| line.unwrap() + "\n")
+        .collect::<String>();
+    let result_from_string = backend.string_to_circuit(&unparsed_file);
+    assert!(result_from_string.is_ok());
 }